+++ modules/php-apache-mysql/jefe-cli.sh
@@ -9,43 +9,8 @@ load_containers_names(){
     APP_CONTAINER_NAME="${project_name}_php"
 }
 
-# Docker compose var env configuration.
-docker_env() {
-    puts "Docker compose var env configuration." BLUE
-    #     if [[ ! -f "$PROYECT_DIR/.env" ]]; then
-    #         cp $PROYECT_DIR/default.env $PROYECT_DIR/.env
-    #     fi
-    echo "" > $PROYECT_DIR/.env
-    set_dotenv PROJECT_TYPE $project_type
-    puts "Write project name (default $project_type):" MAGENTA
-    read proyect_name
-    if [ -z $proyect_name ]; then
-        set_dotenv PROJECT_NAME $project_type
-        proyect_name=$project_type
-    else
-        set_dotenv PROJECT_NAME $proyect_name
-    fi
-    puts "Write project root, directory path from your proyect (default src):" MAGENTA
-    read option
-    if [ -z $option ]; then
-        set_dotenv PROJECT_ROOT "../src"
-    else
-        set_dotenv PROJECT_ROOT "../$option"
-    fi
-    puts "Write vhost (default $proyect_name.local):" MAGENTA
-    read option
-    if [ -z $option ]; then
-        set_dotenv VHOST "$proyect_name.local"
-    else
-        set_dotenv VHOST $option
-    fi
-    puts "Write environment var name, (default development):" MAGENTA
-    read option
-    if [ -z $option ]; then
-        set_dotenv ENVIRONMENT "development"
-    else
-        set_dotenv ENVIRONMENT "$option"
-    fi
+# Configure environments vars of module for docker image.
+module_docker_env() {
     puts "Write database name (default $proyect_name):" MAGENTA
     read option
     if [ -z $option ]; then
@@ -119,12 +84,11 @@ docker_env() {
     set_dotenv FRAMEWORK $framework
     puts "Database root password is password" YELLOW
     set_dotenv DB_ROOT_PASSWORD "password"
-    puts "phpMyAdmin url: phpmyadmin.$vhost" YELLOW
 }
 
 # Add vhost of /etc/hosts file
-set_vhost(){
-    remove_vhost # Remove old vhost
+set-vhost(){
+    remove-vhost # Remove old vhost
     if [ ! "$( grep jefe-cli_wordpress /etc/hosts )" ]; then
         puts "Setting vhost..." BLUE
         load_dotenv
@@ -151,14 +115,6 @@ permissions(){
 
 # Create dump of the database of the proyect.
 dump() {
-    usage= cat <<EOF
-dump [-e] [--environment] [-f] [--file] [-h] [--help]
-
-Arguments:
-    -e, --environment		Set environment to import dump. Default is docker
-    -f, --file			File name of dump. Default is dump.sql
-    -h, --help			Print Help (this message) and exit
-EOF
     # set an initial value
     ENVIRONMENT="docker"
     FILE_NAME="dump.sql"
@@ -173,7 +129,7 @@ EOF
         case "$1" in
             -e|--environment) ENVIRONMENT=$2 ; shift 2 ;;
             -f|--file) FILE_NAME=$2 ; shift 2 ;;
-            -h|--help) echo $usage ; exit 1 ; shift ;;
+            -h|--help) usage_dump ; exit 1 ; shift ;;
             --) shift ; break ;;
             *) echo "Internal error!" ; exit 1 ;;
         esac
@@ -189,15 +145,7 @@ EOF
 }
 
 # Import dump of dumps folder of the proyect.
-import_dump() {
-    usage= cat <<EOF
-import_dump [-e] [--environment] [-f] [--file] [-h] [--help]
-
-Arguments:
-    -e, --environment		Set environment to import dump. Default is docker
-    -f, --file			File name of dump to import. Defualt is dump.sql
-    -h, --help			Print Help (this message) and exit
-EOF
+import-dump() {
     # set an initial value for the flag
     ENVIRONMENT="docker"
     FILE_NAME="dump.sql"
@@ -212,7 +160,7 @@ EOF
         case "$1" in
             -e|--environment) ENVIRONMENT=$2 ; shift 2 ;;
             -f|--file) FILE_NAME=$2 ; shift 2 ;;
-            -h|--help) echo $usage ; exit 1 ; shift ;;
+            -h|--help) usage_import_dump ; exit 1 ; shift ;;
             --) shift ; break ;;
             *) echo "Internal error!" ; exit 1 ;;
         esac
@@ -229,13 +177,6 @@ EOF
 
 # Delete database and create empty database.
 resetdb() {
-    usage= cat <<EOF
-resetdb [-e] [--environment] [-h] [--help]
-
-Arguments:
-    -e, --environment		Set environment to import dump. Default is docker
-    -h, --help			Print Help (this message) and exit
-EOF
     # set an initial value for the flag
     ENVIRONMENT="docker"
 
@@ -248,7 +189,7 @@ EOF
     while true ; do
         case "$1" in
             -e|--environment) ENVIRONMENT=$2 ; shift 2 ;;
-            -h|--help) echo $usage ; exit 1 ; shift ;;
+            -h|--help) usage_resetdb ; exit 1 ; shift ;;
             --) shift ; break ;;
             *) echo "Internal error!" ; exit 1 ;;
         esac
@@ -263,7 +204,7 @@ EOF
     fi
 }
 
-composer_install() {
+composer-install() {
     e=$1
     if [ -z "${e}" ]; then
         e="docker"
@@ -277,7 +218,7 @@ composer_install() {
     fi
 }
 
-composer_update() {
+composer-update() {
     e=$1
     if [ -z "${e}" ]; then
         e="docker"