+++ modules/php-nginx-mysql/jefe-cli.sh
@@ -11,43 +11,8 @@ load_containers_names(){
     NGINX_CONTAINER_NAME="${project_name}_nginx"
 }
 
-# Docker compose var env configuration.
-docker_env() {
-    puts "Docker compose var env configuration." BLUE
-    #     if [[ ! -f "$PROYECT_DIR/.env" ]]; then
-    #         cp $PROYECT_DIR/default.env $PROYECT_DIR/.env
-    #     fi
-    echo "" > $PROYECT_DIR/.env
-    set_dotenv PROJECT_TYPE $project_type
-    puts "Write project name (default $project_type):" MAGENTA
-    read proyect_name
-    if [ -z $proyect_name ]; then
-        set_dotenv PROJECT_NAME $project_type
-        proyect_name=$project_type
-    else
-        set_dotenv PROJECT_NAME $proyect_name
-    fi
-    puts "Write project root, directory path from your proyect (default src):" MAGENTA
-    read option
-    if [ -z $option ]; then
-        set_dotenv PROJECT_ROOT "../src"
-    else
-        set_dotenv PROJECT_ROOT "../$option"
-    fi
-    puts "Write vhost (default $proyect_name.local):" MAGENTA
-    read option
-    if [ -z $option ]; then
-        set_dotenv VHOST "$proyect_name.local"
-    else
-        set_dotenv VHOST $option
-    fi
-    puts "Write environment var name, (default development):" MAGENTA
-    read option
-    if [ -z $option ]; then
-        set_dotenv ENVIRONMENT "development"
-    else
-        set_dotenv ENVIRONMENT "$option"
-    fi
+# Configure environments vars of module for docker image.
+module_docker_env() {
     puts "Write database name (default $proyect_name):" MAGENTA
     read option
     if [ -z $option ]; then
@@ -155,14 +120,6 @@ after_up(){
 
 # Create dump of the database of the proyect.
 dump() {
-    usage= cat <<EOF
-dump [-e] [--environment] [-f] [--file] [-h] [--help]
-
-Arguments:
-    -e, --environment		Set environment to import dump. Default is docker
-    -f, --file			File name of dump. Default is dump.sql
-    -h, --help			Print Help (this message) and exit
-EOF
     # set an initial value for the flag
     ENVIRONMENT="docker"
     FILE_NAME="dump.sql"
@@ -177,7 +134,7 @@ EOF
         case "$1" in
             -e|--environment) ENVIRONMENT=$2 ; shift 2 ;;
             -f|--file) FILE_NAME=$2 ; shift 2 ;;
-            -h|--help) echo $usage ; exit 1 ; shift ;;
+            -h|--help) usage_dump ; exit 1 ; shift ;;
             --) shift ; break ;;
             *) echo "Internal error!" ; exit 1 ;;
         esac
@@ -189,15 +146,7 @@ EOF
 }
 
 # Import dump of dumps folder of the proyect.
-import_dump() {
-    usage= cat <<EOF
-import_dump [-e] [--environment] [-f] [--file] [-h] [--help]
-
-Arguments:
-    -e, --environment		Set environment to import dump. Default is docker
-    -f, --file			File name of dump to import. Defualt is dump.sql
-    -h, --help			Print Help (this message) and exit
-EOF
+import-dump() {
     # set an initial value for the flag
     ENVIRONMENT="docker"
     FILE_NAME="dump.sql"
@@ -212,7 +161,7 @@ EOF
         case "$1" in
             -e|--environment) ENVIRONMENT=$2 ; shift 2 ;;
             -f|--file) FILE_NAME=$2 ; shift 2 ;;
-            -h|--help) echo $usage ; exit 1 ; shift ;;
+            -h|--help) usage_import_dump ; exit 1 ; shift ;;
             --) shift ; break ;;
             *) echo "Internal error!" ; exit 1 ;;
         esac
@@ -225,13 +174,6 @@ EOF
 
 # Delete database and create empty database.
 resetdb() {
-    usage= cat <<EOF
-resetdb [-e] [--environment] [-h] [--help]
-
-Arguments:
-    -e, --environment		Set environment to import dump. Default is docker
-    -h, --help			Print Help (this message) and exit
-EOF
     # set an initial value for the flag
     ENVIRONMENT="docker"
 
@@ -244,7 +186,7 @@ EOF
     while true ; do
         case "$1" in
             -e|--environment) ENVIRONMENT=$2 ; shift 2 ;;
-            -h|--help) echo $usage ; exit 1 ; shift ;;
+            -h|--help) usage_resetdb ; exit 1 ; shift ;;
             --) shift ; break ;;
             *) echo "Internal error!" ; exit 1 ;;
         esac
@@ -259,7 +201,7 @@ EOF
 }
 
 # Execute the command "composer install" in workdir folder
-composer_install() {
+composer-install() {
     e=$1
     if [ -z "${e}" ]; then
         e="docker"
@@ -273,7 +215,7 @@ composer_install() {
 }
 
 # Execute the command "composer update" in workdir folder
-composer_update() {
+composer-update() {
     e=$1
     if [ -z "${e}" ]; then
         e="docker"
@@ -289,16 +231,6 @@ composer_update() {
 if [[ $FRAMEWORK == "Laravel" ]]; then
 # Execute the command "php artisan migrate" in workdir folder. Running laravel migrations
 migrate() {
-        usage= cat <<EOF
-migrate [-e] [--environment] [-f] [--force] [--refresh] [--refresh-seed] [-h] [--help]
-
-Arguments:
-    -e, --environment		Set environment to import dump. Default is docker
-    -f, --force			Force Migrations to run in production (migrate
-        --refresh			Roll back all of your migrations and then execute the  migrate command
-        --refresh-seed			Roll back all of your migrations, execute the  migrate command and run all database seed
-    -h, --help			Print Help (this message) and exit
-EOF
         # set an initial value for the flag
         ENVIRONMENT="docker"
         MIGRATE_OPTION=""
@@ -315,7 +247,7 @@ EOF
                 -f|--force) MIGRATE_OPTION=' --force' ; shift 2 ;;
                 --refresh) MIGRATE_OPTION=':refresh' ; shift 2 ;;
                 --refresh-seed) MIGRATE_OPTION=':refresh --seed' ; shift 2 ;;
-                -h|--help) echo $usage ; exit 1 ; shift ;;
+                -h|--help) usage_migrate ; exit 1 ; shift ;;
                 --) shift ; break ;;
                 *) echo "Internal error!" ; exit 1 ;;
             esac