+++ modules/wordpress/jefe-cli.sh
@@ -3,7 +3,7 @@
 # wordpress jefe-cli.sh
 
 # Load utilities
-source ~/.jefe-cli/libs/utilities.sh
+source $DIR/libs/utilities.sh
 
 # load container names vars
 load_containers_names(){
@@ -12,48 +12,16 @@ load_containers_names(){
     APP_CONTAINER_NAME="${project_name}_wordpress"
 }
 
-# Configure environments vars of docker.
-docker_env() {
-    puts "Docker compose var env configuration." BLUE
-    echo "" > .jefe/.env
-    set_dotenv PROJECT_TYPE $project_type
-    puts "Write project name (default $project_type):" MAGENTA
-    read proyect_name
-    if [ -z $proyect_name ]; then
-        set_dotenv PROJECT_NAME $project_type
-        proyect_name=$project_type
-    else
-        set_dotenv PROJECT_NAME $proyect_name
-    fi
-    puts "Write project root, directory path from your proyect (default src):" MAGENTA
-    read option
-    if [ -z $option ]; then
-        set_dotenv PROJECT_ROOT "../src/"
-    else
-        set_dotenv PROJECT_ROOT "../${option}/"
-    fi
-    puts "Write vhost (default $proyect_name.local):" MAGENTA
-    read option
-    if [ -z $option ]; then
-        set_dotenv VHOST "$proyect_name.local"
-    else
-        set_dotenv VHOST $option
-    fi
-    puts "Write environment var name, (default development):" MAGENTA
-    read option
-    if [ -z $option ]; then
-        set_dotenv ENVIRONMENT "development"
-    else
-        set_dotenv ENVIRONMENT "$option"
-    fi
-    puts "Write wordpress version, (default latest):" MAGENTA
+# Configure environments vars of module for docker image.
+module_docker_env() {
+    puts "Write WordPress version, (default latest):" MAGENTA
     read option
     if [ -z $option ]; then
         set_dotenv WORDPRESS_VERSION "latest"
     else
         set_dotenv WORDPRESS_VERSION "$option"
     fi
-    puts "Write wordpress table prefix (default wp_):" MAGENTA
+    puts "Write WordPress table prefix (default wp_):" MAGENTA
     read option
     if [ -z $option ]; then
         set_dotenv WORDPRESS_TABLE_PREFIX "wp_"
@@ -63,13 +31,12 @@ docker_env() {
 
     puts "Database root password is password" YELLOW
     set_dotenv DB_ROOT_PASSWORD "password"
-    puts "Database name is wordpress" YELLOW
+    puts "Database name is WordPress" YELLOW
     set_dotenv DB_NAME "wordpress"
-    puts "Database user is wordpress" YELLOW
+    puts "Database user is WordPress" YELLOW
     set_dotenv DB_USER "wordpress"
-    puts "Database wordpress password is wordpress" YELLOW
+    puts "Database WordPress password is wordpress" YELLOW
     set_dotenv DB_PASSWORD "wordpress"
-    puts "phpMyAdmin url: phpmyadmin.$vhost" YELLOW
 }
 
 # Fix permisions of the proyect folder
@@ -84,14 +51,6 @@ after_up(){
 
 # Create dump of the database of the proyect.
 dump() {
-    usage= cat <<EOF
-dump [-e] [--environment] [-f] [--file] [-h] [--help]
-
-Arguments:
-    -e, --environment		Set environment to import dump. Default is docker
-    -f, --file			File name of dump. Default is dump.sql
-    -h, --help			Print Help (this message) and exit
-EOF
     # set an initial value for the flag
     ENVIRONMENT="docker"
     FILE_NAME="dump.sql"
@@ -106,7 +65,7 @@ EOF
         case "$1" in
             -e|--environment) ENVIRONMENT=$2 ; shift 2 ;;
             -f|--file) FILE_NAME=$2 ; shift 2 ;;
-            -h|--help) echo $usage ; exit 1 ; shift ;;
+            -h|--help) usage_dump ; exit 1 ; shift ;;
             --) shift ; break ;;
             *) echo "Internal error!" ; exit 1 ;;
         esac
@@ -121,14 +80,7 @@ EOF
 }
 
 # Import dump of dumps folder of the proyect.
-import_dump() {
-    usage= cat <<EOF
-import_dump [-f] [--file] [-h] [--help]
-
-Arguments:
-    -f, --file			File name of dump to import. Defualt is dump.sql
-    -h, --help			Print Help (this message) and exit
-EOF
+import-dump() {
     # set an initial value for the flag
     ENVIRONMENT="docker"
     FILE_NAME="dump.sql"
@@ -143,26 +95,19 @@ EOF
         case "$1" in
             -e|--environment) ENVIRONMENT=$2 ; shift 2 ;;
             -f|--file) FILE_NAME=$2 ; shift 2 ;;
-            -h|--help) echo $usage ; exit 1 ; shift ;;
+            -h|--help) usage_import_dump ; exit 1 ; shift ;;
             --) shift ; break ;;
             *) echo "Internal error!" ; exit 1 ;;
         esac
     done
 
     docker exec -i ${project_name}_db mysql -u ${dbuser} -p"${dbpassword}" ${dbname}  < "./dumps/${FILE_NAME}"
-    set_siteurl
+    set-siteurl
 
 }
 
 # Delete database and create empty database.
 resetdb() {
-    usage= cat <<EOF
-resetdb [-e] [--environment] [-h] [--help]
-
-Arguments:
-    -e, --environment		Set environment to import dump. Default is docker
-    -h, --help			Print Help (this message) and exit
-EOF
     # set an initial value for the flag
     ENVIRONMENT="docker"
 
@@ -175,7 +120,7 @@ EOF
     while true ; do
         case "$1" in
             -e|--environment) ENVIRONMENT=$2 ; shift 2 ;;
-            -h|--help) echo $usage ; exit 1 ; shift ;;
+            -h|--help) usage_resetdb ; exit 1 ; shift ;;
             --) shift ; break ;;
             *) echo "Internal error!" ; exit 1 ;;
         esac
@@ -190,15 +135,7 @@ EOF
 }
 
 # Update siteurl and home options value in wordpress database.
-set_siteurl() {
-    usage= cat <<EOF
-set_siteurl [-e] [--environment] [-H] [--host] [-h] [--help]
-
-Arguments:
-    -e, --environment		Set environment to import dump. Default is docker
-    -H, --host			Host to set. Defualt value of the VHOST configured
-    -h, --help			Print Help (this message) and exit
-EOF
+set-siteurl() {
     # set an initial value for the flag
     ENVIRONMENT="docker"
     HOST="$VHOST"
@@ -213,7 +150,7 @@ EOF
         case "$1" in
             -e|--environment) ENVIRONMENT=$2 ; shift 2 ;;
             -H|--host) HOST=$2 ; shift 2 ;;
-            -h|--help) echo $usage ; exit 1 ; shift ;;
+            -h|--help) usage_set_siteurl ; exit 1 ; shift ;;
             --) shift ; break ;;
             *) echo "Internal error!" ; exit 1 ;;
         esac
@@ -231,14 +168,6 @@ EOF
 
 # Synchronize files to the selected environment.
 deploy() {
-    usage= cat <<EOF
-ps [-e <environment>] [--environment <environment>] [-t] [--test] [-h] [--help]
-
-Arguments:
-    -e, --environment		Set environment to deployed
-    -t, --test			Perform a test of the files to be synchronized
-    -h, --help			Print Help (this message) and exit
-EOF
     # set an initial value for the flag
     ENVIRONMENT=""
     TEST=false
@@ -253,7 +182,7 @@ EOF
         case "$1" in
             -e|--environment) ENVIRONMENT=$2 ; shift 2 ;;
             -t|--test) TEST=true ; shift ;;
-            -h|--help) echo $usage ; exit 1 ; shift ;;
+            -h|--help) usage_deploy ; exit 1 ; shift ;;
             --) shift ; break ;;
             *) echo "Internal error!" ; exit 1 ;;
         esac
@@ -292,7 +221,7 @@ EOF
 }
 
 # Execute the command "composer install" in workdir folder.
-composer_install() {
+composer-install() {
     e=$1
     if [ -z "${e}" ]; then
         e="docker"
@@ -306,7 +235,7 @@ composer_install() {
 }
 
 # Execute the command "composer update" in workdir folder.
-composer_update() {
+composer-update() {
     e=$1
     if [ -z "${e}" ]; then
         e="docker"
@@ -321,14 +250,6 @@ composer_update() {
 
 # Define wordpress debug to true or false.
 debug(){
-    usage= cat <<EOF
-debug [true] [false] [-h] [--help]
-
-Arguments:
-    false			Define wordpress debug to false
-    true			Define wordpress debug to true
-    -h, --help			Print Help (this message) and exit
-EOF
     # set an initial value for the flag
     DEBUG=false
 
@@ -340,7 +261,7 @@ EOF
     # extract options and their arguments into variables.
     while true ; do
         case "$1" in
-            -h|--help) echo $usage ; exit 1 ; shift ;;
+            -h|--help) usage_debug ; exit 1 ; shift ;;
             --) shift ; break ;;
             *) echo "Internal error!" ; exit 1 ;;
         esac
@@ -349,43 +270,14 @@ EOF
     if [[ true == "$1" ]]; then
         DEBUG=true
     fi
-    puts "Setting wordpress debug..."
+    puts "Setting wordpress debug as '$DEBUG'" YELLOW
     docker exec -it $APP_CONTAINER_NAME bash -c "sed -i \"s/define('WP_DEBUG', .*);/define('WP_DEBUG', $DEBUG);/g\" wp-config.php"
 }
 
-php-extensions-install() {
-    docker exec -it $APP_CONTAINER_NAME bash -c "
-        apt-get update && \
-        apt-get install -y libfreetype6-dev libjpeg62-turbo-dev libmcrypt-dev  curl git && \
-        rm -rf /var/lib/apt/lists/* && \
-        docker-php-ext-configure gd --with-freetype-dir=/usr/include/ --with-jpeg-dir=/usr/include/  &&  \
-        docker-php-ext-install intl && \
-        docker-php-ext-install gd && \
-        docker-php-ext-install zip && \
-        docker-php-ext-install mcrypt && \
-        docker-php-ext-install pdo_mysql
-    "
-    # Install composer
-    docker exec -it $APP_CONTAINER_NAME bash -c "
-        php -r \"copy('https://getcomposer.org/installer', 'composer-setup.php');\" && \
-        php -r \"if (hash_file('SHA384', 'composer-setup.php') === '544e09ee996cdf60ece3804abc52599c22b1f40f4323403c44d44fdfdd586475ca9813a858088ffbc1f233e9b180f061') { echo 'Installer verified'; } else { echo 'Installer corrupt'; unlink('composer-setup.php'); } echo PHP_EOL;\" && \
-        php composer-setup.php && \
-        php -r \"unlink('composer-setup.php');\" && \
-        mv composer.phar /usr/local/bin/composer
-    "
-    # Install composer
-    docker exec -it $APP_CONTAINER_NAME bash -c " echo '
-log_errors = On
-error_log = /dev/stderr
-short_open_tag=On
-max_execution_timen = 3000
-max_input_time = 3000
-memory_limit = 1000
-post_max_size = 100M
-upload_max_filesize = 100M
-' > /usr/local/etc/php/conf.d/php.ini
-    "
-#     docker exec -it $APP_CONTAINER_NAME bash -c "apt-get update && apt-get install -y zlib1g-dev && rm -rf /var/lib/apt/lists/* && docker-php-ext-install zip"
+# Generate tab completion strings.
+module_completions() {
+    completions="composer-install composer-update set-siteurl debug"
+    echo $completions
 }
 
 # Initialice